Move Clock binders to separate file
